pyproject.toml
[tool.poetry]
name = "core-tutorials"
version = "0.1.0"
description = ""
authors = ["Shah Zeb"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
django-filter = "^24.2"
django-split-settings = "^1.3.2"
pyyaml = "^6.0.1"
editorconfig = "^0.12.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
